backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
pypandoc
langchain
openai
tiktoken
diff-match-patch
beautifulsoup4
